// src/app/shared/interfaces/product-characteristics.interface.ts
export interface IProdCharacter {
    // bodyMaterial: { name: string, char: Array<string> };
    // plafondMaterial: Array<string>;
    // bodyColor: Array<string>;
    // plafondColor: Array<string>;
    // height: Array<string>;
    // allHeight: Array<string>;
    // length: Array<string>;
    // width: Array<string>;
    // diameter: Array<string>;
    // outside: Array<string>;
    // typeOfLightSource: Array<string>;
    // lampBase: Array<string>;
    // ledServiceLife: Array<string>;
    // lampCount: Array<string>;
    // lampsIncluded: Array<string>;
    // lampPower: Array<string>;
    // luminousFlux: Array<string>;
    // dimer: Array<string>;
    // remoteСontrol: Array<string>;
    // degreeOfProtection: Array<string>;
    // power: Array<string>;
    // powerSupply: Array<string>;
    id: string;
    name: string;
    select: Array<string>;
}